refactor: dead code removal, schema consolidation, and service split
- Fix s3_sync_service bug: handle_upload_complete missing image_id arg
- Remove dead code: sagemaker_repository.py, ConfirmationBadge.tsx,
  unused date utils, get_s3_object_bytes
- Remove unused packages: opencv-python-headless, python-magic, uuid
- Consolidate inline BaseModel schemas into schemas/ modules
- Refactor extraction_service to template method pattern
- Merge duplicated OCR processing methods into _process_ocr_page
- Split image listing/deletion from upload_service into image_list_service
- Update .gitignore for planning docs and temp artifacts
