+++ lambda/api/app/schemas/__init__.py
@@ -7,6 +7,8 @@ from .schema import SchemaGenerateRequest, SchemaSaveRequest
 from .job import JobStartResponse
 from .image import ImageInfo
 from .app import CustomPromptRequest
+from .image_operations import ProcessRequest, VerificationRequest, SuggestionStatusUpdate, S3ImportRequest
+from .usecase import UsecaseToolsUpdate
 
 __all__ = [
     # OCR
@@ -27,4 +29,11 @@ __all__ = [
     "ImageInfo",
     # App
     "CustomPromptRequest",
+    # Image Operations
+    "ProcessRequest",
+    "VerificationRequest",
+    "SuggestionStatusUpdate",
+    "S3ImportRequest",
+    # Usecase
+    "UsecaseToolsUpdate",
 ]
