+++ lambda/api/app/schemas/usecase.py
@@ -1,0 +1,6 @@
+"""Schemas for usecase-related requests"""
+from pydantic import BaseModel
+
+
+class UsecaseToolsUpdate(BaseModel):
+    tool_ids: list[str]
