+++ lambda/api/app/services/upload_service.py
@@ -5,24 +5,18 @@ from datetime import datetime
 from typing import Dict, Any, Optional
 
 from repositories import (
-    create_image_record, get_image, get_images, update_image_status, update_converted_image,
-    get_children_by_parent_id, delete_image as repo_delete_image
+    create_image_record, get_image, update_image_status, update_converted_image,
 )
 from schemas import (
     PresignedUrlRequest, PresignedUrlResponse, UploadCompleteRequest
 )
 from config import settings
-from utils import resize_image, decimal_to_float
+from utils import resize_image
 from repositories import get_app_schemas, get_app_input_methods
-from repositories.usecase_repository import get_permitted_app_names
-from repositories import user_repository
 from background import BackgroundTaskExtension
 from services.pdf_conversion_service import convert_pdf_to_image
-from schemas.image import ImageInfo
 
 logger = logging.getLogger(__name__)
-
-# 共通のS3クライアントを使用
 
 
 class UploadService:
@@ -76,8 +70,8 @@ class UploadService:
                 filename=request.filename,
                 s3_key=s3_key,
                 app_name=request.app_name,
-                status="uploading",  # アップロード中ステータスを設定
-                page_processing_mode=request.page_processing_mode,  # 追加
+                status="uploading",
+                page_processing_mode=request.page_processing_mode,
                 uploaded_by=uploaded_by
             )
 
@@ -183,7 +177,6 @@ class UploadService:
                     "resize_image function not available, skipping resize")
         except Exception as e:
             logger.error(f"画像リサイズエラー: {str(e)}")
-            # リサイズに失敗しても処理を続行
 
     async def _handle_pdf_conversion(self, image_id: str, request: UploadCompleteRequest) -> Dict[str, Any]:
         """PDF変換処理"""
@@ -235,12 +228,10 @@ class UploadService:
 
             # 使用するS3キーを決定
             if converted_s3_keys:
-                # 変換後の画像がある場合
                 target_s3_keys = converted_s3_keys
                 bucket_name = self.bucket_name
                 logger.info(f"変換後の画像のダウンロードURLを生成します: {bucket_name}")
             elif s3_keys:
-                # 元画像を使用
                 target_s3_keys = s3_keys
                 bucket_name = self.bucket_name
                 logger.info(f"元画像のダウンロードURLを生成します: {bucket_name}")
@@ -297,7 +288,7 @@ class UploadService:
             logger.info(f"Generated download URL for image {image_id}")
 
             return {
-                "presigned_url": main_presigned_url,  # 単一画像用のメインURL
+                "presigned_url": main_presigned_url,
                 "presigned_urls": presigned_urls,
                 "total_pages": len(presigned_urls),
                 "is_multipage": len(presigned_urls) > 1,
@@ -308,163 +299,4 @@ class UploadService:
 
         except Exception as e:
             logger.error(f"Error generating download URL: {str(e)}")
-            raise
-
-    @staticmethod
-    def _serialize_images(images: list[dict]) -> list[dict]:
-        """DynamoDB の画像レコードを API レスポンス形式（camelCase）に変換する"""
-        result = []
-        for img in images:
-            try:
-                # DynamoDB の Decimal 型を Python の int/float に変換
-                converted = decimal_to_float(img)
-                info = ImageInfo.model_validate(converted)
-                result.append(info.model_dump(by_alias=True))
-            except Exception as e:
-                logger.error(f"Image serialization error for {img.get('id', '?')}: {e}; raw_keys={sorted(img.keys())}")
-                result.append({"id": img.get("id", ""), "name": img.get("filename", ""), "status": img.get("status", "")})
-        return result
-
-    async def get_images_list(self, app_name: str = None, uploaded_by: str = None) -> Dict[str, Any]:
-        """画像一覧を取得する"""
-        try:
-            images = get_images(app_name, uploaded_by=uploaded_by)
-            self._enrich_uploaded_by_email(images)
-
-            serialized = self._serialize_images(images)
-            result = {
-                "images": serialized,
-                "total": len(serialized)
-            }
-
-            logger.info(f"Retrieved {len(serialized)} images")
-            return result
-
-        except Exception as e:
-            logger.error(f"Error getting images list: {str(e)}")
-            raise
-
-    async def get_images_for_user(self, user_id: str, role: str, app_name: str = None) -> Dict[str, Any]:
-        """ユーザーの権限に応じた画像一覧を取得する
-
-        Args:
-            user_id: ユーザーID
-            role: システムロール（admin / author / reader）
-            app_name: ユースケースでフィルタする場合に指定
-        """
-        if role == "admin":
-            return await self.get_images_list(app_name)
-
-        permitted = get_permitted_app_names(user_id)
-        if not permitted:
-            return {"images": [], "total": 0}
-
-        return await self.get_images_for_permitted_apps(permitted, app_name_filter=app_name)
-
-    async def get_images_for_permitted_apps(self, app_names: list[str], app_name_filter: str = None) -> Dict[str, Any]:
-        """権限のあるユースケースの画像一覧を取得する
-
-        Args:
-            app_names: ユーザーが権限を持つ app_name のリスト
-            app_name_filter: 特定のユースケースでさらに絞り込む場合に指定
-        """
-        try:
-            if app_name_filter:
-                # フィルタ指定時は権限チェック済みの app_name のみ取得
-                if app_name_filter not in app_names:
-                    return {"images": [], "total": 0}
-                target_apps = [app_name_filter]
-            else:
-                target_apps = app_names
-
-            all_images = []
-            for name in target_apps:
-                images = get_images(app_name=name)
-                all_images.extend(images)
-
-            # upload_time 降順でソート
-            all_images.sort(key=lambda x: x.get("upload_time", ""), reverse=True)
-            self._enrich_uploaded_by_email(all_images)
-
-            serialized = self._serialize_images(all_images)
-            return {
-                "images": serialized,
-                "total": len(serialized)
-            }
-        except Exception as e:
-            logger.error(f"Error getting images for permitted apps: {str(e)}")
-            raise
-
-    @staticmethod
-    def _enrich_uploaded_by_email(images: list[dict]) -> None:
-        """画像リストに uploaded_by_email / verified_by_email を付与する"""
-        subs = set()
-        for img in images:
-            if img.get("uploaded_by"):
-                subs.add(img["uploaded_by"])
-            if img.get("verified_by"):
-                subs.add(img["verified_by"])
-        if not subs:
-            return
-        email_map = user_repository.get_emails_by_cognito_subs(subs)
-        for img in images:
-            img["uploaded_by_email"] = email_map.get(img.get("uploaded_by", ""), "")
-            img["verified_by_email"] = email_map.get(img.get("verified_by", ""), "")
-
-    async def delete_image(self, image_id: str, cognito_sub: str = None, is_admin: bool = False) -> Dict[str, Any]:
-        """画像を削除する
-
-        Args:
-            image_id: 削除対象の画像ID
-            cognito_sub: 操作ユーザーの cognito_sub（所有者チェック用）
-            is_admin: admin ロールの場合 True（所有者チェックをスキップ）
-        """
-        try:
-            image = get_image(image_id)
-            if not image:
-                raise ValueError("Image not found")
-
-            # 所有者チェック（admin 以外）
-            if not is_admin:
-                if not cognito_sub or image.get("uploaded_by") != cognito_sub:
-                    raise PermissionError("Forbidden: not the owner")
-            
-            parent_document_id = image.get("parent_document_id")
-            page_processing_mode = image.get("page_processing_mode")
-            total_pages = image.get("total_pages", 0)
-            
-            # 親ファイルの場合（個別処理で2ページ以上、parent_document_idなし）
-            is_parent = (not parent_document_id and 
-                        page_processing_mode == "individual" and 
-                        total_pages > 1)
-            
-            if is_parent:
-                # 親ファイルの場合、全ての子ファイルも削除
-                children = get_children_by_parent_id(image_id)
-                for child in children:
-                    repo_delete_image(child['id'])
-                    logger.info(f"Deleted child image: {child['id']}")
-            
-            # 子ファイルの場合、削除前に残りの子ファイル数をチェック
-            remaining_count = 0
-            if parent_document_id:
-                all_children = get_children_by_parent_id(parent_document_id)
-                # 削除前の子ファイル数をカウント（自分自身を除く）
-                remaining_count = len([c for c in all_children if c['id'] != image_id])
-                logger.info(f"Remaining children count (before deletion): {remaining_count}")
-            
-            # 対象ファイルを削除
-            repo_delete_image(image_id)
-            logger.info(f"Deleted image: {image_id}")
-            
-            # 子ファイルの場合、残りが0なら親も削除
-            if parent_document_id and remaining_count == 0:
-                # 子ファイルが全て削除されたら親も削除
-                repo_delete_image(parent_document_id)
-                logger.info(f"Deleted parent image: {parent_document_id}")
-            
-            return {"status": "success", "message": "Image deleted successfully"}
-
-        except Exception as e:
-            logger.error(f"Error deleting image: {str(e)}")
             raise
