+++ web/package.json
@@ -15,14 +15,12 @@
     "lucide-react": "^0.577.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.22.3",
-    "uuid": "^11.0.5"
+    "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
     "@types/node": "^22.10.10",
     "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.22",
-    "@types/uuid": "^9.0.8",
     "@vitejs/plugin-react": "^4.2.1",
     "autoprefixer": "^10.4.20",
     "postcss": "^8.5.1",
